+++ src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/common/server/BUILD.bazel
@@ -5,23 +5,28 @@ load("//src/main/docker:macros.bzl", "java_image")
 kt_jvm_library(
     name = "internal_reporting_server",
     srcs = ["InternalReportingServer.kt"],
-    visibility = [
-        "//src/main/kotlin/org/wfanet/measurement/integration/common/reporting:__subpackages__",
-        "//src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2:__subpackages__",
-    ],
     deps = [
-        "//src/main/proto/wfa/measurement/internal/reporting/v2:basic_reports_service_kt_jvm_grpc_proto",
-        "//src/main/proto/wfa/measurement/internal/reporting/v2:measurement_consumers_service_kt_jvm_grpc_proto",
-        "//src/main/proto/wfa/measurement/internal/reporting/v2:measurements_service_kt_jvm_grpc_proto",
-        "//src/main/proto/wfa/measurement/internal/reporting/v2:metric_calculation_specs_service_kt_jvm_grpc_proto",
-        "//src/main/proto/wfa/measurement/internal/reporting/v2:metrics_service_kt_jvm_grpc_proto",
-        "//src/main/proto/wfa/measurement/internal/reporting/v2:report_schedule_iterations_service_kt_jvm_grpc_proto",
-        "//src/main/proto/wfa/measurement/internal/reporting/v2:report_schedules_service_kt_jvm_grpc_proto",
-        "//src/main/proto/wfa/measurement/internal/reporting/v2:reporting_sets_service_kt_jvm_grpc_proto",
-        "//src/main/proto/wfa/measurement/internal/reporting/v2:reports_service_kt_jvm_grpc_proto",
-        "@wfa_common_jvm//imports/kotlin/kotlinx/coroutines:core",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/grpc",
+        "//src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/common/server:internal_reporting_server",
+        "//src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/common/service:services",
+        "@wfa_common_jvm//imports/java/picocli",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/db/postgres:flags",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/db/r2dbc/postgres",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/gcloud/spanner",
     ],
+)
+
+java_binary(
+    name = "InternalReportingServer",
+    main_class = "org.wfanet.measurement.reporting.deploy.v2.common.server.InternalReportingServerKt",
+    runtime_deps = [":internal_reporting_server"],
+)
+
+java_image(
+    name = "internal_reporting_server_image",
+    binary = ":InternalReportingServer",
+    main_class = "org.wfanet.measurement.reporting.deploy.v2.common.server.InternalReportingServerKt",
+    visibility = ["//src:docker_image_deployment"],
 )
 
 kt_jvm_library(
