+++ src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/gcloud/server/BUILD.bazel
@@ -3,11 +3,11 @@ load("@wfa_rules_kotlin_jvm//kotlin:defs.bzl", "kt_jvm_library")
 load("//src/main/docker:macros.bzl", "java_image")
 
 kt_jvm_library(
-    name = "gcloud_postgres_internal_reporting_server",
-    srcs = ["GCloudPostgresInternalReportingServer.kt"],
+    name = "gcloud_internal_reporting_server",
+    srcs = ["GCloudInternalReportingServer.kt"],
     runtime_deps = ["@wfa_common_jvm//imports/java/com/google/cloud/sql/postgres:r2dbc"],
     deps = [
-        "//src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/common/server:internal_reporting_server",
+        "//src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/common/server:abstract_internal_reporting_server",
         "//src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/common/service:services",
         "@wfa_common_jvm//imports/java/io/r2dbc",
         "@wfa_common_jvm//imports/java/picocli",
@@ -20,17 +20,17 @@ kt_jvm_library(
 )
 
 java_binary(
-    name = "GCloudPostgresInternalReportingServer",
-    main_class = "org.wfanet.measurement.reporting.deploy.v2.gcloud.postgres.server.GCloudPostgresInternalReportingServerKt",
+    name = "GCloudInternalReportingServer",
+    main_class = "org.wfanet.measurement.reporting.deploy.v2.gcloud.server.GCloudInternalReportingServerKt",
     runtime_deps = [
-        ":gcloud_postgres_internal_reporting_server",
+        ":gcloud_internal_reporting_server",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/gcloud/logging",
     ],
 )
 
 java_image(
-    name = "gcloud_postgres_internal_reporting_server_image",
-    binary = ":GCloudPostgresInternalReportingServer",
-    main_class = "org.wfanet.measurement.reporting.deploy.v2.gcloud.postgres.server.GCloudPostgresInternalReportingServerKt",
+    name = "gcloud_internal_reporting_server_image",
+    binary = ":GCloudInternalReportingServer",
+    main_class = "org.wfanet.measurement.reporting.deploy.v2.gcloud.server.GCloudInternalReportingServerKt",
     visibility = ["//src:docker_image_deployment"],
 )
