--- src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/postgres/server/BUILD.bazel
@@ -1,30 +1,0 @@
-load("@rules_java//java:defs.bzl", "java_binary")
-load("@wfa_rules_kotlin_jvm//kotlin:defs.bzl", "kt_jvm_library")
-load("//src/main/docker:macros.bzl", "java_image")
-
-kt_jvm_library(
-    name = "postgres_internal_reporting_server",
-    srcs = ["PostgresInternalReportingServer.kt"],
-    deps = [
-        "//src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/common/server:internal_reporting_server",
-        "//src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/common/service:services",
-        "@wfa_common_jvm//imports/java/picocli",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/db/postgres:flags",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/db/r2dbc/postgres",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/gcloud/spanner",
-    ],
-)
-
-java_binary(
-    name = "PostgresInternalReportingServer",
-    main_class = "org.wfanet.measurement.reporting.deploy.v2.postgres.server.PostgresInternalReportingServerKt",
-    runtime_deps = [":postgres_internal_reporting_server"],
-)
-
-java_image(
-    name = "postgres_internal_reporting_server_image",
-    binary = ":PostgresInternalReportingServer",
-    main_class = "org.wfanet.measurement.reporting.deploy.v2.postgres.server.PostgresInternalReportingServerKt",
-    visibility = ["//src:docker_image_deployment"],
-)
